Remove unused declarations and fix potential overflow in MLX backend
Remove 8 unused function declarations from layers.h that had no
implementations (EncoderLayer, PolicyHeadClassical, PolicyHeadConvolution,
PolicyHeadAttention, ValueHead, MovesLeftHead, PositionEncoding,
DynamicPositionEncoding). These operations are implemented inline in
network_mlx.cc.

Use size_t instead of int for index calculations in
ConvertConvWeightsOIHWtoOHWI() to prevent potential integer overflow
with large tensors.

+++ src/neural/backends/mlx/layers.cc
@@ -57,8 +57,8 @@ mx::array ConvertConvWeightsOIHWtoOHWI(const std::vector<float>& weights,
     for (int ic = 0; ic < inChannels; ic++) {
       for (int h = 0; h < kH; h++) {
         for (int w = 0; w < kW; w++) {
-          int srcIdx = ((oc * inChannels + ic) * kH + h) * kW + w;
-          int dstIdx = ((oc * kH + h) * kW + w) * inChannels + ic;
+          size_t srcIdx = ((static_cast<size_t>(oc) * inChannels + ic) * kH + h) * kW + w;
+          size_t dstIdx = ((static_cast<size_t>(oc) * kH + h) * kW + w) * inChannels + ic;
           converted[dstIdx] = weights[srcIdx];
         }
       }
